Window resizing bug fix and improved audio thumbnail
Fixed a bug causing the application to crash when resizing the window. Also started working on improving the audio thumbnail so that it will respond to changes in the audio buffer. Currently it responds to changing the value of the gain slider, however this only works on short audio files at the moment,